# apps/web/openapi.json
{
  "components": {
    "schemas": {
...
      "FsDeleteReq": {
        "properties": {
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
      "FsDeleteRes": {
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
      "FsListItem": {
        "properties": {
          "full_path": {
            "type": "string"
          },
          "is_dir": {
            "type": "boolean"
          },
          "last_modified": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "size": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "name",
          "full_path",
          "is_dir",
          "size",
          "last_modified"
        ],
        "type": "object"
      },
      "FsListReq": {
        "properties": {
          "parent_path": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "FsReadTextReq": {
        "properties": {
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
      "FsReadTextRes": {
        "properties": {
          "content": {
            "type": "string"
          }
        },
        "required": [
          "content"
        ],
        "type": "object"
      },
...
      "MediaRootCreateReq": {
        "description": "新增媒体资源根目录请求。",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
      "MediaRootRow": {
        "description": "媒体资源根目录列表行。",
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "last_scanned_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "path",
          "name",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "MediaSubtitleRow": {
        "description": "视频拥有的字幕文件列表行。",
        "properties": {
          "file_name": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "file_size": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "modified_at": {
            "type": "string"
          },
          "scanned_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "file_name",
          "file_path",
          "file_size",
          "modified_at",
          "scanned_at"
        ],
        "type": "object"
      },
      "MediaVideoDeleteReq": {
        "description": "删除媒体库视频请求。",
        "properties": {
          "video_paths": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "video_paths"
        ],
        "type": "object"
      },
      "MediaVideoDeleteRes": {
        "description": "删除媒体库视频结果摘要。",
        "properties": {
          "deleted_subtitles": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "deleted_videos": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "deleted_videos",
          "deleted_subtitles"
        ],
        "type": "object"
      },
      "MediaVideoRow": {
        "description": "媒体库视频列表行。",
        "properties": {
          "file_name": {
            "type": "string"
          },
          "file_path": {
            "type": "string"
          },
          "file_size": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "modified_at": {
            "type": "string"
          },
          "root_id": {
            "format": "int64",
            "type": "integer"
          },
          "scanned_at": {
            "type": "string"
          },
          "subtitle_count": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "subtitles": {
            "items": {
              "$ref": "#/components/schemas/MediaSubtitleRow"
            },
            "type": "array"
          }
        },
        "required": [
          "id",
          "root_id",
          "file_name",
          "file_path",
          "file_size",
          "modified_at",
          "scanned_at",
          "subtitle_count",
          "subtitles"
        ],
        "type": "object"
      },
      "MediaVideosPageRes": {
        "description": "媒体库视频分页查询结果。",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/MediaVideoRow"
            },
            "type": "array"
          },
          "total": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total"
        ],
        "type": "object"
      },
...
      "VideoPlaybackProbeRes": {
        "properties": {
          "audio_codec": {
            "type": [
              "string",
              "null"
            ]
          },
          "container": {
            "type": [
              "string",
              "null"
            ]
          },
          "direct_playable": {
            "description": "是否建议走直链（H.264/AAC 等浏览器常见组合）",
            "type": "boolean"
          },
          "needs_transcode": {
            "description": "是否建议服务端转码后再播",
            "type": "boolean"
          },
          "video_codec": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "direct_playable",
          "needs_transcode"
        ],
        "type": "object"
      },
      "VideoTranscodePhase": {
        "enum": [
          "idle",
          "running",
          "ready",
          "failed"
        ],
        "type": "string"
      },
      "VideoTranscodeStatusRes": {
        "properties": {
          "message": {
            "type": [
              "string",
              "null"
            ]
          },
          "phase": {
            "$ref": "#/components/schemas/VideoTranscodePhase"
          },
          "progress": {
            "description": "0.0–1.0，仅 `running` 时有效",
            "format": "float",
            "type": [
              "number",
              "null"
            ]
          }
        },
        "required": [
          "phase"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "name": ""
    },
    "title": "media-admin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/fs/delete-subtitle": {
      "post": {
        "operationId": "deleteSubtitleFs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsDeleteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FsDeleteRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "fs"
        ]
      }
    },
    "/api/fs/list": {
      "post": {
        "operationId": "listFs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsListReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FsListItem"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "fs"
        ]
      }
    },
    "/api/fs/read-text": {
      "post": {
        "operationId": "readTextFs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FsReadTextReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FsReadTextRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "fs"
        ]
      }
    },
    "/api/fs/video/probe": {
      "get": {
        "operationId": "probeVideoFs",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoPlaybackProbeRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "fs"
        ]
      }
    },
    "/api/fs/video/transcode/start": {
      "post": {
        "operationId": "startVideoTranscodeFs",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTranscodeStatusRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "fs"
        ]
      }
    },
    "/api/fs/video/transcode/status": {
      "get": {
        "operationId": "videoTranscodeStatusFs",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoTranscodeStatusRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "fs"
        ]
      }
    },
...
    "/api/media-library/files": {
      "get": {
        "operationId": "listFilesMediaLibrary",
        "parameters": [
          {
            "in": "query",
            "name": "root_id",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "has_subtitle",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "current",
            "required": false,
            "schema": {
              "format": "int32",
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "format": "int32",
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaVideosPageRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "media-library"
        ]
      }
    },
    "/api/media-library/roots": {
      "get": {
        "operationId": "listRootsMediaLibrary",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MediaRootRow"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "media-library"
        ]
      },
      "post": {
        "operationId": "createRootMediaLibrary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaRootCreateReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaRootRow"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "media-library"
        ]
      }
    },
    "/api/media-library/roots/{id}": {
      "delete": {
        "operationId": "deleteRootMediaLibrary",
        "parameters": [
          {
            "description": "媒体资源目录 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "media-library"
        ]
      }
    },
    "/api/media-library/roots/{id}/scan": {
      "post": {
        "operationId": "scanRootMediaLibrary",
        "parameters": [
          {
            "description": "媒体资源目录 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "media-library"
        ]
      }
    },
    "/api/media-library/videos/delete": {
      "post": {
        "operationId": "deleteVideosMediaLibrary",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MediaVideoDeleteReq"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaVideoDeleteRes"
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "media-library"
        ]
      }
    },
...
  },
...
}